+++ NEWS.md
@@ -1,5 +1,5 @@
-Version 0.8.5-dev
-2012-02-??
+Version 0.8.5
+2012-02-11
 
  - Do not assume any thrown exception is an Error (#65)
  - Clear both aut and non-auth render cache on style change (#59)